Changed internal block table structures. Compressed block tables and hash tabels should now load. Also, block table offsets > 4GB should load fine now…

+++ trunk/CrystalMpq/CrystalMpq/SeedNotFoundException.cs
@@ -16,7 +16,7 @@ namespace CrystalMpq
 	/// <remarks>The seed is needed for reading encrypted files.</remarks>
 	public sealed class SeedNotFoundException : Exception
 	{
-		internal SeedNotFoundException(int block)
+		internal SeedNotFoundException(long block)
 			: base(string.Format(ErrorMessages.GetString("SeedNotFound"), block)) { }
 	}
 }
